fastapi_backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
